Change the type of LocalSession::track_start to int
Track start is the vertical poisition of the track in pixesls
Int is enough - maximum is a lot of kilometers

+++ cinelerra/localsession.h
@@ -79,7 +79,7 @@ public:
 	ptstime loop_start;
 	ptstime loop_end;
 // Vertical start of track view
-	int64_t track_start;
+	int track_start;
 // Horizontal start of view in seconds
 	ptstime view_start_pts;
 // Zooming of the timeline. Length of media per pixel
